feat(auth): implementar hook useAuthStatus para gerenciar estado de autenticação
